removed some warnings

+++ Application/Source/MeshBuilder.cpp
@@ -88,7 +88,7 @@ Mesh* MeshBuilder::Generatesphere(const std::string& meshName, float lengthX, fl
 			v.color.Set(name); 
 			v.normal.Set(cos((phi + 10) * Rad) * cos(theta * Rad),  sin((phi + 10) * Rad), cos((phi + 10) * Rad) * sin(theta * Rad));
 			x = float((cos((theta/2) * Rad)/2.8)+0.5);
-			y = (sin((phi + 10) * Rad) / 2)+0.5;
+			y = (sin((phi + 10) * Rad) / 2)+0.5f;
 			v.texCoord.Set(x, y);
 		
 			vertex_buffer_data.push_back(v);
@@ -122,11 +122,11 @@ Mesh* MeshBuilder::GenerateTorus(const std::string& meshName, unsigned numStack,
 	Mesh* mesh = new Mesh(meshName);
 	for (unsigned stack = 0; stack < numStack + 1; stack++) {
 		for (unsigned slice = 0; slice < numSlice + 1; slice++) {
-			z1 = outerR * cos(stack * degreePerStack*Rad);
-			x1 = outerR * sin(stack * degreePerStack*Rad);
-			z2 = (outerR + innerR * cos(slice * degreePerSlice*Rad)) * cos(stack * degreePerStack*Rad);
-			y2 = innerR * sin(slice * degreePerSlice*Rad);
-			x2 = (outerR + innerR * cos(float(slice) * degreePerSlice*Rad)) * sin(stack * degreePerStack*Rad);
+			z1 = outerR * cos(float(stack) * degreePerStack*Rad);
+			x1 = outerR * sin(float(stack) * degreePerStack*Rad);
+			z2 = (outerR + innerR * cos(float(slice) * degreePerSlice*Rad)) * cos(float(stack) * degreePerStack*Rad);
+			y2 = innerR * sin(float(slice) * degreePerSlice*Rad);
+			x2 = (outerR + innerR * cos(float(slice) * degreePerSlice*Rad)) * sin(float(stack) * degreePerStack*Rad);
 			v.pos.Set(x2, y2, z2);
 			v.color.Set(color);
 			v.normal.Set(x2-x1,y2,z2-z1);
@@ -431,14 +431,14 @@ Mesh* MeshBuilder::GenerateColliderBox(const std::string& meshName, double legnt
 	v.pos.Set(float(-legnthx / 2), float(lengthy / 2), float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
 	v.pos.Set(float(legnthx / 2), float(lengthy / 2), float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
 	//sides
-	v.pos.Set(float(-legnthx) / 2, lengthy / 2, -lengthz / 2); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(-legnthx) / 2, -lengthy / 2, -lengthz / 2); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(legnthx) / 2, lengthy / 2, -lengthz / 2); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(legnthx) / 2, -lengthy / 2, -lengthz / 2); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(-legnthx) / 2, lengthy / 2, lengthz / 2); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(-legnthx) / 2, -lengthy / 2, lengthz / 2); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(legnthx) / 2, lengthy / 2, lengthz / 2); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(legnthx) / 2, -lengthy / 2, lengthz / 2); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(-legnthx) / 2, float(lengthy / 2), float(-lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(-legnthx) / 2, float(-lengthy / 2), float(-lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(legnthx) / 2, lengthy / 2, float(-lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(legnthx) / 2, -lengthy / 2, float(-lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(-legnthx) / 2, lengthy / 2, float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(-legnthx) / 2, -lengthy / 2, float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(legnthx) / 2, lengthy / 2, float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(legnthx) / 2, -lengthy / 2, float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
 	Mesh* mesh = new Mesh(meshName);
 	for (int i = 0; i < 24; i++) {
 		index_buffer_data.push_back(i);
